obnk/requirements.txt
Django==1.11.7
django-cors-headers==2.1.0
django-extensions==1.9.7
django-filter==1.0.4
djangorestframework==3.7.3
djangorestframework-camel-case==0.2.0
gunicorn==19.7.1
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.2
pytz==2017.3
six==1.11.0
typing==3.6.2